Moved the filterMember test
